# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feedback & Review Analyzer API",
    "description": "REST API that queues review-file analysis jobs and serves dashboard analytics.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analytics/summary": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Analytics Summary",
        "operationId": "analytics_summary_api_v1_analytics_summary_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Job UUID",
              "title": "Job Id"
            },
            "description": "Job UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsSummary"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/analytics/sentiment-breakdown": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Sentiment Breakdown",
        "operationId": "sentiment_breakdown_api_v1_analytics_sentiment_breakdown_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentBreakdownResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/analytics/trends": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Trends",
        "operationId": "trends_api_v1_analytics_trends_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "weekly or monthly",
              "title": "Interval"
            },
            "description": "weekly or monthly"
          },
          {
            "name": "theme_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Theme Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendsResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyticsSummary": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/SummaryBlock"
          },
          "sentiment_breakdown": {
            "$ref": "#/components/schemas/SentimentBreakdown"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/DashboardTheme"
            },
            "type": "array",
            "title": "Themes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "sentiment_breakdown",
          "themes"
        ],
        "title": "AnalyticsSummary",
        "example": {
          "sentiment_breakdown": {
            "negative": 55,
            "neutral": 20,
            "positive": 25
          },
          "summary": {
            "drift_status": "Low (0.03)",
            "model_validation_accuracy": "78.0%",
            "net_sentiment_score": -18,
            "overall_sentiment": "Negative Trend",
            "pii_redacted_count": 1420,
            "total_reviews": 10000
          },
          "themes": [
            {
              "count": 3420,
              "general_class": "Billing & Subscriptions",
              "id": "theme-1",
              "name": "Charged Twice Monthly Subscription",
              "sample_verbatims": [
                {
                  "id": "v12",
                  "sentiment": "Negative",
                  "text": "Charged twice for my monthly sub. Support email [EMAIL_REDACTED] hasn't replied."
                }
              ],
              "sentiment": "Strongly Negative",
              "sentiment_score": -0.82
            }
          ]
        }
      },
...
      "DashboardTheme": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "general_class": {
            "type": "string",
            "title": "General Class"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "sentiment": {
            "type": "string",
            "title": "Sentiment"
          },
          "sentiment_score": {
            "type": "number",
            "title": "Sentiment Score"
          },
          "sample_verbatims": {
            "items": {
              "$ref": "#/components/schemas/VerbatimSample"
            },
            "type": "array",
            "title": "Sample Verbatims"
          }
        },
        "type": "object",
        "required": [
          "id",
          "general_class",
          "name",
          "count",
          "sentiment",
          "sentiment_score"
        ],
        "title": "DashboardTheme"
      },
...
      "ErrorBody": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorBody",
        "example": {
          "code": "INVALID_FILE_FORMAT",
          "details": {},
          "error": "Invalid file format. Only .csv and .xlsx supported."
        }
      },
...
      "SentimentBreakdown": {
        "properties": {
          "positive": {
            "type": "integer",
            "title": "Positive"
          },
          "neutral": {
            "type": "integer",
            "title": "Neutral"
          },
          "negative": {
            "type": "integer",
            "title": "Negative"
          }
        },
        "type": "object",
        "required": [
          "positive",
          "neutral",
          "negative"
        ],
        "title": "SentimentBreakdown"
      },
      "SentimentBreakdownResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "sentiment_breakdown": {
            "$ref": "#/components/schemas/SentimentBreakdown"
          },
          "counts": {
            "$ref": "#/components/schemas/SentimentCounts"
          },
          "total_reviews": {
            "type": "integer",
            "title": "Total Reviews"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "sentiment_breakdown",
          "counts",
          "total_reviews"
        ],
        "title": "SentimentBreakdownResponse",
        "example": {
          "counts": {
            "negative": 3800,
            "neutral": 2000,
            "positive": 2500,
            "strongly_negative": 1700
          },
          "job_id": "3f1c8a2e-4b91-4c77-9d12-7a8e0b1c2d3e",
          "sentiment_breakdown": {
            "negative": 55,
            "neutral": 20,
            "positive": 25
          },
          "total_reviews": 10000
        }
      },
      "SentimentCounts": {
        "properties": {
          "positive": {
            "type": "integer",
            "title": "Positive"
          },
          "neutral": {
            "type": "integer",
            "title": "Neutral"
          },
          "negative": {
            "type": "integer",
            "title": "Negative"
          },
          "strongly_negative": {
            "type": "integer",
            "title": "Strongly Negative"
          }
        },
        "type": "object",
        "required": [
          "positive",
          "neutral",
          "negative",
          "strongly_negative"
        ],
        "title": "SentimentCounts"
      },
      "SummaryBlock": {
        "properties": {
          "total_reviews": {
            "type": "integer",
            "title": "Total Reviews"
          },
          "overall_sentiment": {
            "type": "string",
            "title": "Overall Sentiment"
          },
          "net_sentiment_score": {
            "type": "number",
            "title": "Net Sentiment Score"
          },
          "pii_redacted_count": {
            "type": "integer",
            "title": "Pii Redacted Count"
          },
          "model_validation_accuracy": {
            "type": "string",
            "title": "Model Validation Accuracy"
          },
          "drift_status": {
            "type": "string",
            "title": "Drift Status"
          }
        },
        "type": "object",
        "required": [
          "total_reviews",
          "overall_sentiment",
          "net_sentiment_score",
          "pii_redacted_count",
          "model_validation_accuracy",
          "drift_status"
        ],
        "title": "SummaryBlock"
      },
...
      "TrendPoint": {
        "properties": {
          "period_start": {
            "type": "string",
            "title": "Period Start"
          },
          "period_label": {
            "type": "string",
            "title": "Period Label"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          },
          "average_sentiment_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Sentiment Score"
          },
          "net_sentiment_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Sentiment Score"
          },
          "positive": {
            "type": "integer",
            "title": "Positive"
          },
          "neutral": {
            "type": "integer",
            "title": "Neutral"
          },
          "negative": {
            "type": "integer",
            "title": "Negative"
          },
          "low_sample": {
            "type": "boolean",
            "title": "Low Sample"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_label",
          "review_count",
          "positive",
          "neutral",
          "negative",
          "low_sample"
        ],
        "title": "TrendPoint"
      },
      "TrendsResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "interval": {
            "type": "string",
            "title": "Interval"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "excluded_without_date": {
            "type": "integer",
            "title": "Excluded Without Date"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "interval",
          "excluded_without_date",
          "data"
        ],
        "title": "TrendsResponse",
        "example": {
          "data": [
            {
              "average_sentiment_score": -0.21,
              "low_sample": false,
              "negative": 57,
              "net_sentiment_score": -17.5,
              "neutral": 31,
              "period_label": "2026-W28",
              "period_start": "2026-07-06",
              "positive": 24,
              "review_count": 112
            }
          ],
          "excluded_without_date": 12,
          "interval": "weekly",
          "job_id": "3f1c8a2e-4b91-4c77-9d12-7a8e0b1c2d3e",
          "timezone": "UTC"
        }
      },
...
      "VerbatimSample": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "sentiment": {
            "type": "string",
            "title": "Sentiment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "sentiment"
        ],
        "title": "VerbatimSample",
        "example": {
          "id": "v12",
          "sentiment": "Negative",
          "text": "Charged twice for my monthly sub. Support email [EMAIL_REDACTED] hasn't replied."
        }
      },
...
    }
  }
}